game/pyproject.toml
[tool.poetry]
name = "online game"
version = "0.1.0"
description = ""
authors = ["helisson victor"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pygame = "^2.6.1"
websockets = "^13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
